Move ATInternetTracking into components
