+++ .prettierignore
@@ -2,6 +2,6 @@ dist
 .dist
 storybook-static
 site/source/locales/*.yaml
-site/source/ATInternetTracking/smarttag.js
+site/source/components/ATInternetTracking/smarttag.js
 *.yaml.d.ts
 .yarnrc.yml
