+++ site/source/components/Feedback/index.tsx
@@ -3,7 +3,7 @@ import { Trans, useTranslation } from 'react-i18next'
 import { useLocation } from 'react-router-dom'
 import styled from 'styled-components'
 
-import { TrackingContext } from '@/ATInternetTracking'
+import { TrackingContext } from '@/components/ATInternetTracking'
 import { Popover } from '@/design-system'
 import { Button } from '@/design-system/buttons'
 import { Emoji } from '@/design-system/emoji'
