+++ site/source/components/layout/Footer/Privacy.tsx
@@ -6,8 +6,8 @@ import { Checkbox } from '@/design-system/field'
 import { Link } from '@/design-system/typography/link'
 import { Body, SmallBody } from '@/design-system/typography/paragraphs'
 
-import { TrackPage, TrackingContext } from '../../../ATInternetTracking'
 import * as safeLocalStorage from '../../../storage/safeLocalStorage'
+import { TrackPage, TrackingContext } from '../../ATInternetTracking'
 
 export default function Privacy({
 	label,
