+++ site/source/pages/Budget/Budget.tsx
@@ -3,6 +3,7 @@ import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
+import { TrackPage } from '@/components/ATInternetTracking'
 import MoreInfosOnUs from '@/components/MoreInfosOnUs'
 import { ScrollToTop } from '@/components/utils/Scroll'
 import { Markdown } from '@/components/utils/markdown'
@@ -12,7 +13,6 @@ import { Grid } from '@/design-system/layout'
 import { H1, H2 } from '@/design-system/typography/heading'
 import { Body } from '@/design-system/typography/paragraphs'
 
-import { TrackPage } from '../../ATInternetTracking'
 import Meta from '../../components/utils/Meta'
 import prose from './budget.md?raw'
 import budget from './budget.yaml'
