+++ site/source/pages/Creer/GuideStatut/MinorityDirector.tsx
@@ -4,14 +4,13 @@ import {
 	directorIsInAMinority,
 	useDispatchAndGoToNextQuestion,
 } from '@/actions/companyStatusActions'
+import { TrackPage } from '@/components/ATInternetTracking'
 import DefaultHelmet from '@/components/utils/DefaultHelmet'
 import AnswerGroup from '@/design-system/answer-group'
 import { Button } from '@/design-system/buttons'
 import { H2 } from '@/design-system/typography/heading'
 import { Li, Ul } from '@/design-system/typography/list'
 import { Body } from '@/design-system/typography/paragraphs'
-
-import { TrackPage } from '../../../ATInternetTracking'
 
 export default function MinorityDirector() {
 	const { t } = useTranslation()
