+++ site/source/pages/Creer/GuideStatut/index.tsx
@@ -4,13 +4,13 @@ import { useDispatch, useSelector } from 'react-redux'
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom'
 
 import { resetCompanyStatusChoice } from '@/actions/companyStatusActions'
+import { TrackChapter } from '@/components/ATInternetTracking'
 import { FromBottom } from '@/components/ui/animate'
 import { H1 } from '@/design-system/typography/heading'
 import { Link } from '@/design-system/typography/link'
 import { RootState } from '@/reducers/rootReducer'
 import { useSitePaths } from '@/sitePaths'
 
-import { TrackChapter } from '../../../ATInternetTracking'
 import AutoEntrepreneur from './AutoEntrepreneur'
 import DirectorStatus from './DirectorStatus'
 import MinorityDirector from './MinorityDirector'
