+++ site/source/pages/Nouveautes/Nouveautes.tsx
@@ -3,6 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { Navigate, useMatch, useNavigate } from 'react-router-dom'
 import styled from 'styled-components'
 
+import { TrackPage } from '@/components/ATInternetTracking'
 import MoreInfosOnUs from '@/components/MoreInfosOnUs'
 import { determinant, useHideNewsBanner } from '@/components/layout/NewsBanner'
 import Meta from '@/components/utils/Meta'
@@ -19,8 +20,6 @@ import {
 import { Body } from '@/design-system/typography/paragraphs'
 import { useFetchData } from '@/hooks/useFetchData'
 import { useSitePaths } from '@/sitePaths'
-
-import { TrackPage } from '../../ATInternetTracking'
 
 const slugify = (name: string) => name.toLowerCase().replace(' ', '-')
