+++ site/source/pages/Stats/LazyStats.tsx
@@ -1,6 +1,7 @@
 import { Suspense, lazy } from 'react'
 import { useTranslation } from 'react-i18next'
 
+import { TrackPage } from '@/components/ATInternetTracking'
 import MoreInfosOnUs from '@/components/MoreInfosOnUs'
 import PageHeader from '@/components/PageHeader'
 import Privacy from '@/components/layout/Footer/Privacy'
@@ -9,7 +10,6 @@ import { ScrollToTop } from '@/components/utils/Scroll'
 import { Emoji } from '@/design-system/emoji'
 import { Intro } from '@/design-system/typography/paragraphs'
 
-import { TrackPage } from '../../ATInternetTracking'
 import illustrationSvg from './illustration.svg'
 
 const Stats = lazy(() => import('./Stats'))
