+++ site/source/pages/gerer/declaration-charges-sociales-independant/index.tsx
@@ -6,6 +6,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import styled from 'styled-components'
 
 import { updateSituation } from '@/actions/actions'
+import { TrackPage } from '@/components/ATInternetTracking'
 import { Condition, WhenAlreadyDefined } from '@/components/EngineValue'
 import PageHeader from '@/components/PageHeader'
 import RuleInput from '@/components/conversation/RuleInput'
@@ -22,7 +23,6 @@ import { SimulationConfig } from '@/reducers/rootReducer'
 import { situationSelector } from '@/selectors/simulationSelectors'
 import { useSitePaths } from '@/sitePaths'
 
-import { TrackPage } from '../../../ATInternetTracking'
 import { SimpleField, SubSection } from '../components/Fields'
 import { ExplicationsResultatFiscal } from './components/ExplicationResultatFiscal'
 import ResultatsSimples from './components/RésultatSimple'
