+++ site/source/pages/gerer/declaration-revenu-independants/components/DéclarationRevenu.tsx
@@ -5,8 +5,8 @@ import { useTranslation } from 'react-i18next'
 import { useDispatch } from 'react-redux'
 import styled, { css } from 'styled-components'
 
-import { TrackPage } from '@/ATInternetTracking'
 import { updateSituation } from '@/actions/actions'
+import { TrackPage } from '@/components/ATInternetTracking'
 import Value, { Condition } from '@/components/EngineValue'
 import ShareOrSaveSimulationBanner from '@/components/ShareSimulationBanner'
 import { ExplicableRule } from '@/components/conversation/Explicable'
