+++ site/source/pages/gerer/demande-mobilité/EndBlock.tsx
@@ -4,6 +4,7 @@ import { Suspense, lazy, useContext, useRef, useState } from 'react'
 import SignaturePad from 'react-signature-pad-wrapper'
 import { ThemeContext } from 'styled-components'
 
+import { TrackPage, TrackingContext } from '@/components/ATInternetTracking'
 import { Condition } from '@/components/EngineValue'
 import { EngineContext, EngineProvider } from '@/components/utils/EngineContext'
 import { Message, PopoverWithTrigger } from '@/design-system'
@@ -17,7 +18,6 @@ import { Link } from '@/design-system/typography/link'
 import { Li, Ul } from '@/design-system/typography/list'
 import { Body, Intro, SmallBody } from '@/design-system/typography/paragraphs'
 
-import { TrackPage, TrackingContext } from '../../../ATInternetTracking'
 import PDFDocument from './PDFDocument'
 
 const IS_TOUCH_DEVICE = isOnTouchDevice()
