+++ site/source/pages/gerer/demande-mobilité/index.tsx
@@ -3,6 +3,7 @@ import Engine, { PublicodesExpression } from 'publicodes'
 import { Fragment, Suspense, lazy, useCallback, useContext } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 
+import { TrackPage } from '@/components/ATInternetTracking'
 import { WhenApplicable, WhenNotApplicable } from '@/components/EngineValue'
 import RuleInput from '@/components/conversation/RuleInput'
 import BrowserOnly from '@/components/utils/BrowserOnly'
@@ -25,7 +26,6 @@ import {
 	omit,
 } from '@/utils'
 
-import { TrackPage } from '../../../ATInternetTracking'
 import formulaire from './demande-mobilité.yaml'
 
 const LazyEndBlock = import.meta.env.SSR
