+++ site/source/pages/integration/index.tsx
@@ -1,6 +1,7 @@
 import { Trans, useTranslation } from 'react-i18next'
 import { Route, Routes, useLocation } from 'react-router-dom'
 
+import { TrackChapter } from '@/components/ATInternetTracking'
 import { ScrollToTop } from '@/components/utils/Scroll'
 import { Banner, InnerBanner } from '@/design-system/banner'
 import { Emoji } from '@/design-system/emoji'
@@ -8,7 +9,6 @@ import { Link } from '@/design-system/typography/link'
 import { useFetchData } from '@/hooks/useFetchData'
 import { useSitePaths } from '@/sitePaths'
 
-import { TrackChapter } from '../../ATInternetTracking'
 import API from './API'
 import Iframe from './Iframe'
 import Library from './Library'
